+++ sdk/src/main/java/software/amazon/lambda/durable/StepContext.java
@@ -2,6 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 package software.amazon.lambda.durable;
 
+import software.amazon.lambda.durable.context.BaseContext;
+
 public interface StepContext extends BaseContext {
     /** Returns the current retry attempt number (0-based). */
     int getAttempt();
