+++ sdk/src/main/java/software/amazon/lambda/durable/retry/RetryStrategies.java
@@ -3,7 +3,7 @@
 package software.amazon.lambda.durable.retry;
 
 import java.time.Duration;
-import software.amazon.lambda.durable.validation.ParameterValidator;
+import software.amazon.lambda.durable.util.ParameterValidator;
 
 /**
  * Factory class for creating common retry strategies.
